--- app/types/turndown.d.ts
@@ -1,7 +1,0 @@
-// turndown.d.ts
-declare module 'turndown' {
-  export default class TurndownService {
-    constructor();
-    turndown(html: string): string;
-  }
-}
